fix(mobile): wait for live-query release before resync cleanup
Adding the current user to a channel triggers a membership resync, which
cleanup()s and rebuilds the channel-scoped collections. The resync effect ran
that cleanup immediately after flipping to the spinner — but unmounting a screen
does not release its live queries synchronously (they stay registered as
collection dependents until GC). So cleanup() tore a source out from under a
still-live dependent, logging "… was manually cleaned up while live query depends
on it" (though the new channel still rendered).

The sign-out path already guards this with waitForLiveQueryRelease(); the resync
path did not. Extract that helper to collections/live-query-release.ts and await
it in the resync effect before resyncProjectCollections(), mirroring sign-out.

Mobile typecheck unchanged (zero new errors). Unit tests pass.
